refactor: postlistの分離のリファクタリング
